Pulled dev branch, create MenuBar, added .h files
