+++ UI/imgui.ini
@@ -24,9 +24,10 @@ Collapsed=0
 DockId=0x00000004,0
 
 [Window][Hierarchy]
-Pos=243,187
-Size=316,648
+Pos=8,27
+Size=316,685
 Collapsed=0
+DockId=0x00000007,0
 
 [Window][Instector]
 Pos=986,8
@@ -35,26 +36,28 @@ Collapsed=0
 DockId=0x0000000A,0
 
 [Window][Browser]
-Pos=8,8
-Size=1264,704
+Pos=326,541
+Size=708,171
 Collapsed=0
 DockId=0x00000020,0
 
 [Window][Viewport]
-Pos=617,173
-Size=708,527
+Pos=326,27
+Size=708,512
 Collapsed=0
+DockId=0x00000021,0
 
 [Window][Console]
-Pos=8,8
-Size=1264,704
+Pos=326,541
+Size=708,171
 Collapsed=0
 DockId=0x00000020,1
 
 [Window][Inspector]
-Pos=649,107
-Size=315,648
+Pos=1036,27
+Size=236,685
 Collapsed=0
+DockId=0x0000001E,0
 
 [Window][MainDockSpace]
 Pos=60,60
@@ -62,8 +65,8 @@ Size=32,32
 Collapsed=0
 
 [Window][DockSpace Demo]
-Pos=0,0
-Size=1280,720
+Pos=0,19
+Size=1280,701
 Collapsed=0
 
 [Docking][Data]
@@ -94,7 +97,13 @@ DockSpace                           ID=0x77733C41 Pos=8,8 Size=1264,704 Split=X 
               DockNode              ID=0x0000000A Parent=0x00000002 SizeRef=286,463 Selected=0xEEE83228
           DockNode                  ID=0x00000004 Parent=0x0000000B SizeRef=269,773 Selected=0x11AA6B68
         DockNode                    ID=0x0000000C Parent=0x00000014 SizeRef=412,704 Selected=0xE7039252
-DockSpace                           ID=0xFA06BC56 Window=0x4647B76E Pos=8,8 Size=1264,704 Split=Y
-  DockNode                          ID=0x0000001F Parent=0xFA06BC56 SizeRef=0,48
-  DockNode                          ID=0x00000020 Parent=0xFA06BC56 SizeRef=0,16 Selected=0x49278EEE
+DockSpace                           ID=0xFA06BC56 Window=0x4647B76E Pos=8,27 Size=1264,685 Split=X
+  DockNode                          ID=0x00000007 Parent=0xFA06BC56 SizeRef=16,0 Selected=0x29EABFBD
+  DockNode                          ID=0x00000008 Parent=0xFA06BC56 SizeRef=48,0 Split=X
+    DockNode                        ID=0x0000001D Parent=0x00000008 SizeRef=48,0 Split=Y
+      DockNode                      ID=0x0000001F Parent=0x0000001D SizeRef=0,48 Split=Y
+        DockNode                    ID=0x00000021 Parent=0x0000001F SizeRef=0,16 Selected=0x13926F0B
+        DockNode                    ID=0x00000022 Parent=0x0000001F SizeRef=0,48
+      DockNode                      ID=0x00000020 Parent=0x0000001D SizeRef=0,16 Selected=0x49278EEE
+    DockNode                        ID=0x0000001E Parent=0x00000008 SizeRef=16,0 Selected=0xE7039252
